Fix speed unit + i16 overflow + stale 104-byte doc
Telemetry.lua: GIANTS' vehicle.lastSpeed is meters-per-millisecond
(so *1000 = m/s), but the mod did *1000/3600, giving m/s/3.6. That
miscalibrated every speed-driven effect in the daemon (centering
spring ramp, rumble period, lateral-force smoothing). Prefer the
public getLastSpeed() getter (km/h) and fall back to lastSpeed*1000
for objects that don't expose it.

ffb.rs apply(): (lvl - last_constant).abs() is i16 arithmetic and
wraps when both operands straddle zero (e.g. last=+32000, lvl=-32000).
Widen to i32 before the delta check.

IPCWriter.lua / telemetry.rs: header comments still said "104 bytes"
after the 108-byte fix landed; sync the docs.

Telemetry.lua: drop the unused isFront local (never read; the branch
keys on positionZ directly).

+++ daemon/src/ffb.rs
@@ -122,7 +122,8 @@ impl FfbDevice {
     pub fn apply(&mut self, out: &EffectOutput) -> Result<()> {
         if let Some(id) = self.id_constant {
             let lvl = scale_to_i16(out.constant);
-            if (lvl - self.last_constant).abs() > 150 {
+            // Widen to i32 before subtracting: i16::MIN - i16::MAX wraps.
+            if (lvl as i32 - self.last_constant as i32).abs() > 150 {
                 self.update_constant(id, lvl, 0xFFFF)?;
                 self.start(id)?;
                 self.last_constant = lvl;
